Factorisation de FieldDate dans son propre component
