+++ packages/app/src/views/InformationsSimulation/InformationsSimulationForm.tsx
@@ -12,28 +12,20 @@ import {
 import {
   mustBeDate,
   parseTrancheEffectifsFormValue,
-  required
+  required,
+  validateDate
 } from "../../utils/formHelpers";
-import {
-  calendarYear,
-  dateToString,
-  parseDate,
-  Year
-} from "../../utils/helpers";
+import { calendarYear, Year } from "../../utils/helpers";
 
 import ActionBar from "../../components/ActionBar";
 import ActionLink from "../../components/ActionLink";
+import FieldDate from "../../components/FieldDate";
 import FormAutoSave from "../../components/FormAutoSave";
 import FormSubmit from "../../components/FormSubmit";
 import Input, { hasFieldError } from "../../components/Input";
 import RadioLabels from "../../components/RadioLabels";
 import { ButtonSimulatorLink } from "../../components/SimulatorLink";
 import globalStyles from "../../utils/globalStyles";
-
-import DatePicker, { registerLocale } from "react-datepicker";
-import "react-datepicker/dist/react-datepicker.css";
-import fr from "date-fns/locale/fr";
-registerLocale("fr", fr);
 
 ///////////////////
 
@@ -44,19 +36,6 @@ const validate = (value: string) => {
   } else {
     return {
       required: requiredError
-    };
-  }
-};
-
-const validateDate = (value: string) => {
-  const requiredError = required(value);
-  const mustBeDateError = mustBeDate(value);
-  if (!requiredError && !mustBeDateError) {
-    return undefined;
-  } else {
-    return {
-      required: requiredError,
-      mustBeDate: mustBeDateError
     };
   }
 };
@@ -258,56 +237,6 @@ function FieldPeriodeReference({ readOnly }: { readOnly: boolean }) {
   );
 }
 
-const hasMustBeDateError = (meta: FieldMetaState<string>) =>
-  meta.error && meta.touched && meta.error.mustBeDate;
-
-function FieldDate({
-  name,
-  label,
-  readOnly
-}: {
-  name: string;
-  label: string;
-  readOnly: boolean;
-}) {
-  const field = useField(name, { validate: validateDate });
-  const error = hasFieldError(field.meta);
-  const mustBeDateError = hasMustBeDateError(field.meta);
-
-  return (
-    <div css={styles.dateField}>
-      <label
-        css={[styles.label, error && styles.labelError]}
-        htmlFor={field.input.name}
-      >
-        {label}
-      </label>
-      <div css={styles.fieldRow}>
-        <Field name={name} validate={validateDate}>
-          {props => (
-            <DatePicker
-              locale="fr"
-              dateFormat="dd/MM/yyyy"
-              selected={parseDate(props.input.value)}
-              onChange={date =>
-                date ? props.input.onChange(dateToString(date)) : ""
-              }
-              readOnly={readOnly}
-              name={name}
-            />
-          )}
-        </Field>
-      </div>
-      <p css={styles.error}>
-        {error &&
-          (mustBeDateError
-            ? "ce champ doit contenir une date au format jj/mm/aaaa"
-            : "ce champ n’est pas valide, renseignez une date au format jj/mm/aaaa")}
-      </p>
-    </div>
-  );
-}
-
 const styles = {
   container: css({
     display: "flex",
@@ -345,17 +274,6 @@ const styles = {
   dates: css({
     display: "flex",
     justifyContent: "space-between"
-  }),
-  dateField: css({
-    marginTop: 5,
-    input: {
-      display: "flex",
-      fontSize: 14,
-      paddingLeft: 22,
-      paddingRight: 22,
-      height: 38,
-      marginTop: 5
-    }
   }),
   edit: css({
     marginTop: 14,
